Optimize Performance on orderSummary component
We do not want to rerender the orderSummary component if the modal is not shown. To avoid unecessary rerender we use the shouldComponentDidUpdate() to check if the modal is shown if it is shown then should be rerender. We are doing this in modal cause modal component is wrapping the orderSummary component and only if the modal is shown we want to show the orderSummary.