Add global error handling
Better user experience to show any error to the user.
The goal is to create a global error handling to show the error in a modal. We do not want to use the modal in the burgerBuilder component because we want to have a flexible way of showing an error no matter in wich component it occurs. Sow e need to place this wrror handling coponent in high order component higher than burgrBuilder.

So we are creating a high order component which will be able to wrap the BurgerBuildComponent.

After we ctreate the high order component we use interceptors in order to handle the errors. We also use the request interceptor in order to be able to  clear any errors so whenever we sent a request we do not have the error setup anymore.