Fix the bug where we do not get the error in a modal from withErrorHandler component.
The issue here is we cannot see that an error occured throught the get request in BurgerBuildier. That is caused because we are using componentDidMount () in withErrorHandler component when we are creating the interceptors. ComponentDidMount () will work after all ComponentDidMount () of the children component of the withErrorHandler will be executed. The get methof in the child component BurgerBuilder is being executed in componentDidMount. That means that the interseptor will take place after the componentDidMount of the child component is executed. So it seems that we reach out to the web from the burgerbuilder component without setting up the intereceptors yet.

In order to fix that we can use the componentWillMount () instead or better the constructor (ComponentWillMount() will be deprecated)  as they are lifecycle hooks responsible for the  creation life cycle stage of a component.

I will keep componentWillMount for now and change it later with hooks.