Cleanup interceptors when the compoent is not no longer in use
When we wrap a component with the withErrorHandler we create interceptors so if the component that is wrapped up is no longer in use we do not want to use its  old interceptors and  kept them in memory. We remove them by using the componentWillUmnount () which takes place when this instance of the class in withErrorHandlerComponent in not required anymore.