Set the Routes for burgerBuilder and Checkout
Both routes will be loaded because path='/' is treated as a prefix and when checkout will be hit the burgerBilder route will be hit as well as it has path='/'. To prevent this we can add the exact property to the Route with the prefix'/'  or use the Switch Component to wrap the Routes to only load one of the routes the first one which matches the given path. The order of the routes is important only if we remove the exact word.

I used all the solutions just to show that we can use Switch as well.