Pass the ingredients via query Params
We need to encode the ingretients into query parmaeters and then add them to the search prop of the object argument of the push method.

First an array is created where it will include the params. Then looping through the ingredients we are getting the key and values where we add them to the new array. We use the encodeUriComponent which is a help function that  ensures that elements/values can be used in the URL.