Fix bug/ the ingredients are null
Because we were setting the ingredients in componentDidMount(), the child ContactData did not have access yet to them, as compoentDidMount of the parent will
be executed after all of the children componets are rendered and mounted.

We are using here compoentWillMount to fix this, just for now as componentWillMount() is gonna be deprecated.