Pass the props of router to ContactData component
Because we use the render method  on Route for ContactData we loose the props of the route. One way is to add the props to the render method as an argument and destribute them to the ContactData like this '{..}' in order to be able to access them in ContactData component. Another way is to wrap the export of the ContactData component with the 'withRouter'.