Fix/Bug the checkout was not working on refreshing
The checkout was not working on refresh because it could not get soon enough the ingredients. So I added a redirect to home page if ingredients are not there and if they are I am showing the order summary. The logic here is that if there are no ingredients there is no reason to show the order summary.