Redirect user to the checkout page after signin if he already built a burger
There are many ways to do that, you can do that wuth route and add parameters where you keep the information of the ingredients.

Since we have the ingredients stored in redux state I have done this in another way.
The logic is to create another property in state which will store the path that the redirect should have when the after sign in/up sumbission will take place.

After I created the action and the specific case in reducer for updating the path of the authRedirectPath property, I dispatched the action in the burgerBuilder where the button to order is clicked, so the authRedirectPath is updated with the value '/checkout'.

Later on the auth component I added the dynamic path to the Redirect component.