Reset the auth path if user reaches the page of signup/in  whilst not building a burger.
If the user visits the signup/in page without being in the process of building the burger, after he is logged  in he should be redirect to home page and not to checkout page.

In order to do that I used the isBurgerBuilt property of the state which is only true if the user removes or adds an ingredient. If this property is not true and if the the authRedirectPath is not '/' then I dispatch the setAuthRedirectPath action to set the path to '/', to actually reset it.