Fix/bug where user after he logges in he is redirected to homepage isntead of checkout
The user propmpted to loggin before he place the order after logged in was redirected to homepage instead of checkout page. The code of saving the authRedirectPath property of state is in auth component which was previously removed from the routes. Easy fix is to add it again in routes when the user is authenticated.