Fix the bug where the checkout component do not have access to the routing props
The checkout component needs to have access to the routing props that will allow it to get information about the path or load a new page.

In order to achieve this we pass the routing props to the render function as an argument and then forward them to the component.