Convert Modal class based component to Functional component
Use React.memo() to replace shouldComponentUpdate()
React.memo() allows us to optimize performance and only update the component when the props of the component change. Since previously the logic was to optimize the component and re render it only when props `show` and `children` specifically are changed. W can still do that by having a function as a second argument in React.memo() and define when we want the component to be memorized by comparing the previous props with the nextprops