Add a custom error handling hook
I moved the functionality of axios interceptors into a hook. This way can be used with any http Client not just axios. It is also removed from the high order component withErrorHandler and that makes it is reusable as it can be use anywhere we want to get an error from hhtp requests and display it how we want to, maybe not in a modal.